requirements.txt
django==4.0.4
djangorestframework==3.13.1
django_mptt==0.13.4
drf_yasg==1.20.0